Move `StatusSnackbar` to `common` folder
