+++ packages/website/src/common/StatusSnackbar/index.test.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { render } from "@testing-library/react";
+
+import StatusSnackbar from ".";
+
+test("renders as expected", () => {
+  const { container } = render(
+    <StatusSnackbar severity="success" handleClose={jest.fn()} />
+  );
+  expect(container).toMatchSnapshot();
+});
